requirements.txt
asgiref==3.5.2
Django==4.0.1
django-cleanup==5.2.0
django-crispy-forms==1.14.0
Pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.6
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
